API/requirements_faq.txt
PyPDF2==3.0.1
requests==2.31.0
pandas==2.0.3
google-generativeai==0.3.2
python-dotenv==1.0.0